requirements.txt
PyGithub>=2.3.0
openai>=1.30.0
unidiff>=0.7.5
pydantic>=2.7.0
